patch: Add kernel config detect
According to the APatch documentation, CONFIG_KALLSYMS MUST be Enabled in the kernel configuration. However, some devices do not have this feature enabled in their kernels, which will cause patch errors and waste a lot of unnecessary time from developers to answering questions for users.

This commit added kernel configuration detection into KernelPatch to detect kernel configuration and intercept it in the early stage of patching to reduce the workload of answering questions.

Test: Build - Patch kernel - Detected disable - Abort
Signed-off-by: sakurakyuo <sakurakyuo@protonmail.com>