NOT TESTED: Allow patches even when verification fails
Before patching the kernel image, it is possible that the kernel coincidentally contains KP1158\x00\x00, causing the checksum of the later data to fail. This change will only log this error message and allow patching to continue (treating the image as unpatched).

Warning: This change has not been tested.