Allow patches even when verification fails
Before patching the kernel image, it is possible that the kernel coincidentally contains KP1158\x00\x00, causing the checksum of the later data to fail.

This change loops through valid checksums and only determines a patched image if a valid checksum is found. If it is not found, it will be treated as a new image instead of just exiting.